Add boolean to control when run runs in RemoteAPIDownload.
Two commits ago, the while loop in run used True then I switched it to !Thread.currentThread().isInterrupted(), but that prevented the loop from ending when an exception was thrown.  Now, the loop uses a boolean to run and if the thread is interrupted the boolean value is changed to false.