Rm 404 and 403 handling from UpdateBroadcastReceiver.
Because UpdateBroadcastReceiver isn't an activity, it can't easily interact with context. If the URL is incorrect, the update for check issues or for a new apk don't occur.  Whne the user tries to do something in an activity, the 403 or 404 error gets handled there.  When the URL is changed, issues is checked.