Move the downloading alert to BaseActivity and trigger the alert when each activity is created. Remove the downloading alert and downloadAPKFile from Configuration and
This change is necessary because before if a user was in the configuration activity and ignored an update, the code prevented subesquent updates from creating an alerts and from navigating from the activity.  The user could manually click the check update button, but this was not desirable.  Now, if an user is on the configuration page and ignores an update if they navigate away from the configuration activity, subsequent updates are checked for.